### 2019-01-30 #1222 @danielskatz
@matbesancon
It looks like there's an extra } in paper.md in the line:
```\lambda_i \cdot (b_i - Ax_i - By_i) = 0 \forall i \in \{1..m_l\}}.$$```

After you fix it, you can try recompiling by typing ```@whedon generate pdf``` as a new comment in this issue.
