### 2019-01-30 4e85c1b
typo in paper equation
