### 2019-01-30 #1222 @danielskatz
note that the URL for Dempe seems to have some spaces that should be removed as well
