+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @article{dempe2018bilevel,
   title={Bilevel optimization: theory, algorithms and applications},
   author={Dempe, Stephan},
-  journal={Optimization On-line URL http://www. optimization-online. org/DB\_HTML/2018/08/6773. html},
+  journal={Optimization Online http://www.optimization-online.org/DB_FILE/2018/08/6773.pdf},
   year={2018}
 }
 
