### 2019-01-30 #1222 @matbesancon
reviewers suggestions:
```
stsievert
andim
ChrisRackauckas
jakobbossek
sarats
sjvrijn
chkoar
vissarion
malmaud
HaoZeke
taqtiqa-mark
```
(those who seem to be comfortable with optimization and/or Julia packages)
I didn't tag them all because it seemed excessive, I'll let the editors see what seems appropriate
