### 2019-01-31 e30a241
fix equation newlines
