### 2019-01-31 #1222 @matbesancon
sorry for stressing the bot, multiline equations are a bit of a bother
