### 2019-01-31 #1222 @matbesancon
ok now special characters don't appear, this is really not ideal. My kingdom for `\begin{align}`
