### 2019-02-03 #1222 @poulson
@stsievert Would you be willing to review this paper?
