### 2019-02-04 #1222 @jakobbossek
@stsievert Sorry, covered with many other task. Moreover, I am not familiar with the Julia language.
