+++ src/BilevelOptimization.jl
@@ -6,6 +6,7 @@ using JuMP: @variable, @constraint, @objective
 export BilevelLP, build_blp_model, BilevelFlowProblems, VariableType
 
 include("types.jl")
+include("complementarity.jl")
 include("build_model.jl")
 include("bilevel_flows.jl")
 
