+++ src/build_model.jl
@@ -3,7 +3,7 @@
 Build a JuMP model (constraints and objective)
 based on the data from bp and with a solver
 """
-function build_blp_model(bp::BilevelLP, solver)
+function build_blp_model(bp::BilevelLP, solver; comp_method = SOS1Complementarity())
     m = JuMP.Model(solver = solver)
     @variable(m, bp.xl[j] <= x[j=1:bp.nu] <= bp.xu[j])
     @variable(m, y[j=1:bp.nl])
@@ -15,7 +15,7 @@ function build_blp_model(bp::BilevelLP, solver)
         JuMP.setcategory(x[j], :Int)
     end
     # adding SOS1 constraints
-    return build_blp_model(m, bp, x, y)
+    return build_blp_model(m, bp, x, y, comp_method = SOS1Complementarity())
 end
 
 """
@@ -23,7 +23,7 @@ Add the lower-level constraints and optimality conditions to
 an existing JuMP model. This assumes the upper-level feasibility
 constraints and objective have already been set
 """
-function build_blp_model(m::JuMP.Model, bp::BilevelLP, x, y)
+function build_blp_model(m::JuMP.Model, bp::BilevelLP, x, y; comp_method = SOS1Complementarity())
     @variable(m, s[1:bp.ml] >= 0) # lower-level slack variables
     @constraint(m, lowercons[i=1:bp.ml],
         sum(bp.A[i,j]*x[j] for j in 1:bp.nu) +
@@ -40,12 +40,7 @@ function build_blp_model(m::JuMP.Model, bp::BilevelLP, x, y)
         end
     end
     @constraint(m, bp.d .+ bp.F' * x .+ bp.B' * λ  .- σ .== 0.0)
-    for j in Base.OneTo(bp.nl)
-        JuMP.addSOS1(m, [σ[j], y[j]])
-    end
-    for i in Base.OneTo(bp.ml)
-        JuMP.addSOS1(m, [λ[i], s[i]])
-    end
+    add_complementarity_constraint(m, comp_method, s, λ, y, σ)
     return (m, x, y, λ)
 end
 
@@ -53,22 +48,26 @@ end
 Build the bilevel JuMP model from the data without
 grouping everything into a `BilevelLP`
 """
-function build_blp_model(m::JuMP.Model, B::M, d, x, y, s, F) where {M<:MT}
-    ml = size(B)[1]
-    @variable(m, λ[1:ml] >= 0)
-    @constraint(m, d .+ F' * x .+ B' * λ .== 0.0)
-    for i in Base.OneTo(ml)
-        JuMP.addSOS1(m, [λ[i], s[i]])
+function build_blp_model(m::JuMP.Model, B::M, d, x, y, s, F; comp_method = SOS1Complementarity()) where {M<:MT}
+    @variable(m, λ[1:bp.ml] >= 0)
+    @variable(m, σ[1:bp.nl] >= 0.) # dual of lower-level lower bound
+    for j in Base.OneTo(bp.nl)
+        if !bp.yl[j]
+            # sigma at 0 if free variable
+            @constraint(m, σ[j] == 0.)
+        else
+            JuMP.setlowerbound(y[j], 0.)
+        end
     end
+    @constraint(m, d .+ F' * x .+ B' * λ .- σ .== 0.0)
+    add_complementarity_constraint(m, comp_method, s, λ, y, σ)
     return (m, λ)
 end
 
-function build_blp_model(m::JuMP.Model, B::M, d, s) where {M<:MT}
+function build_blp_model(m::JuMP.Model, B::M, d, s; comp_method = SOS1Complementarity()) where {M<:MT}
     (ml,nl) = size(B)
[…]
