### 2019-02-18 81a0a77
additional tests for infeasible bounds & vector
