+++ test/runtests.jl
@@ -272,3 +272,34 @@ end
         end
     end
 end
+
+@testset "Bilevel flow big-M vector bounds" begin
+    bfp = test_bflow()
+    bounds_method = BilevelOptimization.BoundComplementarity(10 .* ones(19), 30.)
+    (m, r, y, f, λ) = build_blp_model(bfp, CbcSolver(), comp_method = bounds_method)
+    st = JuMP.solve(m)
+    @test st === :Optimal
+    @test getobjectivevalue(m) ≈ 6.
+    for j in 1:size(r)[2]
+        for i in 1:size(r)[1]
+            @test getvalue(r[i,j]) ≈ sum(getvalue(y[i,j,:]).*bfp.tax_options[i,j,:]) * getvalue(f[i,j])
+        end
+    end
+    bounds_method = BilevelOptimization.BoundComplementarity(10., 10 .* ones(19))
+    (m, r, y, f, λ) = build_blp_model(bfp, CbcSolver(), comp_method = bounds_method)
+    st = JuMP.solve(m)
+    @test st === :Optimal
+    @test getobjectivevalue(m) ≈ 6.
+    for j in 1:size(r)[2]
+        for i in 1:size(r)[1]
+            @test getvalue(r[i,j]) ≈ sum(getvalue(y[i,j,:]).*bfp.tax_options[i,j,:]) * getvalue(f[i,j])
+        end
+    end
+end
+
+@testset "Bilevel flow big-M infeasible" begin
+    bfp = test_bflow()
+    (m, r, y, f, λ) = build_blp_model(bfp, CbcSolver(), comp_method = BilevelOptimization.BoundComplementarity(0.1, 0.1))
+    st = JuMP.solve(m)
+    @test st === :Infeasible
+end
