### 2019-02-18 b0076cf
exported method types, added doc
