+++ src/BilevelOptimization.jl
@@ -3,7 +3,8 @@ module BilevelOptimization
 import JuMP
 using JuMP: @variable, @constraint, @objective
 
-export BilevelLP, build_blp_model, BilevelFlowProblems, VariableType
+export BilevelLP, build_blp_model, BilevelFlowProblems, VariableType,
+       SOS1Complementarity, BoundComplementarity
 
 include("types.jl")
 include("complementarity.jl")
