+++ Project.toml
@@ -14,3 +14,6 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Test", "Cbc"]
+
+[compat]
+JuMP = "0.18"
