### 2019-02-18 07c48ef
complementarity structure (#5)
* complementarity structure

* fix tests

* fixed method sig

* additional tests for infeasible bounds & vector

* exported method types, added doc

* added upper bound on JuMP

* bound jump in project toml
