+++ test/runtests.jl
@@ -22,7 +22,6 @@ end
 @testset "test basic problem" begin
     bp = test_bp()
     (m, x, y, λ, _) = build_blp_model(bp, CbcSolver())
-    println(λ)
     status = JuMP.solve(m)
     @test status === :Optimal
     xv = JuMP.getvalue(x)
@@ -230,4 +229,77 @@ end
             @test getvalue(r[i,j]) ≈ sum(getvalue(y[i,j,:]).*bfp.tax_options[i,j,:]) * getvalue(f[i,j])
         end
     end
+end
+
+@testset "Complementarity types" begin
+    m = JuMP.Model()
+    # testing non-crashing
+    BilevelOptimization.add_complementarity_constraint(m, SOS1Complementarity, [], [], [], [])
+    BilevelOptimization.add_complementarity_constraint(m, SOS1Complementarity, [], [], [], [])
+end
+
+@testset "Basic problem with specified methods" begin
+    bp = test_bp()
+    (m, x, y, λ) = build_blp_model(bp, CbcSolver(), comp_method = SOS1Complementarity)
+    status = JuMP.solve(m)
+    @test status === :Optimal
+    xv = JuMP.getvalue(x)
+    yv = JuMP.getvalue(y)
+    @test xv[1] ≈ 0.0
+    @test yv[1] ≈ 1.0
+    (m, _, _, _) = build_blp_model(bp, CbcSolver(), comp_method = SOS1Complementarity())
+    status = JuMP.solve(m)
+    @test status === :Optimal
+    # arbitrary big-enough bounds
+    (m, x, y, λ) = build_blp_model(bp, CbcSolver(), comp_method = BoundComplementarity(100.,100.))
+    status = JuMP.solve(m)
+    @test status === :Optimal
+    xv = JuMP.getvalue(x)
+    yv = JuMP.getvalue(y)
+    @test xv[1] ≈ 0.0
+    @test yv[1] ≈ 1.0
+end
+
+@testset "Bilevel flow big-M bounds" begin
+    bfp = test_bflow()
+    (m, r, y, f, λ) = build_blp_model(bfp, CbcSolver(), comp_method = BoundComplementarity(100., 100.))
+    st = JuMP.solve(m)
+    @test st === :Optimal
+    @test getobjectivevalue(m) ≈ 6.
+    for j in 1:size(r)[2]
+        for i in 1:size(r)[1]
+            @test getvalue(r[i,j]) ≈ sum(getvalue(y[i,j,:]).*bfp.tax_options[i,j,:]) * getvalue(f[i,j])
+        end
+    end
+end
+
+@testset "Bilevel flow big-M vector bounds" begin
+    bfp = test_bflow()
+    bounds_method = BoundComplementarity(10 .* ones(19), 30.)
+    (m, r, y, f, λ) = build_blp_model(bfp, CbcSolver(), comp_method = bounds_method)
+    st = JuMP.solve(m)
+    @test st === :Optimal
+    @test getobjectivevalue(m) ≈ 6.
+    for j in 1:size(r)[2]
+        for i in 1:size(r)[1]
+            @test getvalue(r[i,j]) ≈ sum(getvalue(y[i,j,:]).*bfp.tax_options[i,j,:]) * getvalue(f[i,j])
+        end
+    end
+    bounds_method = BoundComplementarity(10., 10 .* ones(19))
+    (m, r, y, f, λ) = build_blp_model(bfp, CbcSolver(), comp_method = bounds_method)
+    st = JuMP.solve(m)
+    @test st === :Optimal
+    @test getobjectivevalue(m) ≈ 6.
+    for j in 1:size(r)[2]
[…]
