+++ src/BilevelOptimization.jl
@@ -3,9 +3,11 @@ module BilevelOptimization
 import JuMP
 using JuMP: @variable, @constraint, @objective
 
-export BilevelLP, build_blp_model, BilevelFlowProblems, VariableType
+export BilevelLP, build_blp_model, BilevelFlowProblems, VariableType,
+       SOS1Complementarity, BoundComplementarity
 
 include("types.jl")
+include("complementarity.jl")
 include("build_model.jl")
 include("bilevel_flows.jl")
 

[…]
