### 2019-02-20 8bf5164
Readme comp (#6)
* add complementarity in readme

* added user-defined complementarity

* change bibtex key
