### 2019-02-21 #1222 @vissarion
@poulson I could help with the review although I have no experience with `Julia`. Is there a deadline?
