### 2019-02-21 #1222 @HaoZeke
@poulson you're welcome, I am much better now, thanks, I would be willing and able to review this package. Julia is one of my favourite languages and optimization problems are also right up my alley.
