### 2019-02-22 a0f96b6
fix greek letters, reduce to 3 pages
