### 2019-02-26 #1222 @poulson
@vissarion Thank you for agreeing to help review this, Vissarion. I think that, given @HaoZeke's Julia expertise, between the two of you, the expertise should be covered. As for a deadline, the soft deadline is typically two weeks.

@HaoZeke I'm glad to hear that you're feeling better, and thank you for your enthusiasm in agreeing to review this paper!

@sjvrijn Hi Sander, it seems that we have this paper covered now, but thank you for the offer to serve as a backup!
