### 2019-02-26 #1278 @whedon
[REVIEW]: A Julia package for bilevel optimization problems
**Submitting author:** @matbesancon (<a href="http://orcid.org/0000-0002-6284-3033">Mathieu Besançon</a>)
**Repository:** <a href="https://github.com/matbesancon/BilevelOptimization.jl" target ="_blank">https://github.com/matbesancon/BilevelOptimization.jl</a>
**Version:** v0.2.1
