### 2019-02-27 618dc1d
added random bilevel generators
