### 2019-02-27 9c26f87
documentation, fixed use of RNG
