### 2019-02-27 #1278 @HaoZeke
Hi, @matbesancon I've initiated the review process and hope to finish in the next couple of days. I will be opening pull requests and issues on the repository tracker, but I might also ask for clarifications here (typically with matters pertaining to the paper itself). In addition, for the `paper.md`, I request that a copy to make changes on [hackmd.io](https://hackmd.io) is made available to me and @vissarion  since it is not appropriate to make changes to it via pull requests.
