### 2019-02-27 matbesancon/BilevelOptimization.jl#7 @HaoZeke
Version Bump [JOSS]
Part of the ongoing JOSS review, https://github.com/openjournals/joss-reviews/issues/1278.
Could you please update the version submitted? [0.1.0 -> 0.1.1 or greater]
