### 2019-02-27 #1278 @HaoZeke
@matbesancon, please update the `paper.md` file to fix the missing DOIs.
