### 2019-02-27 d59737e
readme: Document tests
Part of https://github.com/openjournals/joss-reviews/issues/1278.
