### 2019-02-27 #1278 @matbesancon
The hackmd can be found there:
https://hackmd.io/s/B1e2UUNIE#
