### 2019-02-27 d7655c6
update README, DOI in paper
