### 2019-02-27 #1278 @matbesancon
@HaoZeke the first DOI seems invalid (the preprint does not correspond to the article suggested), I'm fixing the others in https://github.com/matbesancon/BilevelOptimization.jl/pull/9
