### 2019-02-27 08fc9b6
update README, DOI in paper (#9)
