+++ paper/paper.bib
@@ -13,7 +13,8 @@
   number={2},
   pages={295--320},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/15m1020575}
 }
 
 @article{brotcorne2001bilevel,
@@ -24,7 +25,8 @@
   number={4},
   pages={345--358},
   year={2001},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi={10.1287/trsc.35.4.345.10433}
 }
 
 @article{Hayrapetyan2007,
@@ -46,10 +48,14 @@ url="https://doi.org/10.1007/s00446-006-0020-y"
 }
 
 @article{harks2018toll,
-  title={Toll Caps in Privatized Road Networks},
-  author={Harks, Tobias and Schroder, Marc and Vermeulen, Dries},
-  journal={arXiv preprint arXiv:1802.10514},
-  year={2018}
+title = "Toll caps in privatized road networks",
+journal = "European Journal of Operational Research",
+year = "2019",
+issn = "0377-2217",
+doi = {10.1016/j.ejor.2019.01.059},
+author = "Tobias Harks and Marc Schröder and Dries Vermeulen",
+keywords = "Game theory, Competition regulation, Toll caps, Nash equilibrium, Wardrop equilibrium",
+abstract = "We consider a network pricing game on a parallel network with congestion effects in which link owners set tolls for travel so as to maximize profit. A central authority is able to regulate this competition by means of a (uniform) price cap. The first question we want to answer is how such a cap should be designed in order to minimize the total congestion. We provide an algorithm that finds an optimal price cap for networks with affine latency functions and a full support Wardrop equilibrium. Second, we consider the induced network performance at an optimal price cap. We show that for two link networks with affine latency functions, the congestion costs at the optimal price cap are at most 8/7 times the optimal congestion costs. For more general latency functions, this bound goes up to 2 under the assumption that an uncapped Nash equilibrium exists. However, in general such an equilibrium need not exist and this can be used to show that optimal price caps can induce arbitrarily inefficient flows."
 }
 
 @article{pineda19,
