### 2019-02-27 matbesancon/BilevelOptimization.jl#7 @matbesancon
@HaoZeke I just bumped the package to 0.2.0, you mean bumping on https://github.com/openjournals/joss-reviews/issues/1278? I didn't find how to do that
