### 2019-02-28 matbesancon/BilevelOptimization.jl#7 @HaoZeke
I've bumped the version being reviewed and can close this if you confirm that the `paper.md` reflects the changes and features included upto v0.2.0
