### 2019-02-28 matbesancon/BilevelOptimization.jl#7 @matbesancon
yes that's the case
