### 2019-02-28 #1278 @matbesancon
missing DOI still invalid
