### 2019-02-28 matbesancon/BilevelOptimization.jl#7 @HaoZeke
Thanks for the clarification.
