### 2019-02-28 #1278 @HaoZeke
Yeah, we'll just have to let that one go. It does not seem to correspond to any DOI. Even the ISSN number does not seem valid. @matbesancon
