### 2019-02-28 #1278 @matbesancon
yup I just saw it is considered as a book and not article on google scholar now, I'll update the bib
