### 2019-02-28 0360fb3
edit bib for S Dempe
