+++ paper/paper.bib
@@ -1,8 +1,8 @@
-@article{dempe2018bilevel,
+@book{dempe2018bilevel,
   title={Bilevel optimization: theory, algorithms and applications},
   author={Dempe, Stephan},
-  journal={Optimization Online http://www.optimization-online.org/DB_FILE/2018/08/6773.pdf},
-  year={2018}
+  year={2018},
+  publisher={TU Bergakademie Freiberg, Fakult{\"a}t f{\"u}r Mathematik und Informatik}
 }
 
 @article{dunning2017jump,
