### 2019-04-04 matbesancon/BilevelOptimization.jl#14 @vissarion
Documentation [JOSS]
This issue is raised as part of https://github.com/openjournals/joss-reviews/issues/1278

It seems that the documentation of the package is the README file. As far as I understand, the main function is `build_blp_model`. However, first this is not stated in the documentation and second it is poorly documented, actually only the `comp_method` argument is discussed. Also what about other functions, e.g. `flow_constraint_standard`?

It would be useful also to include (either in the documentation or in the joss paper) a complete example of formulating and solving a toll-setting problem as a bilevel optimization problem.
