### 2019-04-04 #1278 @vissarion
Hello, and again sorry for the late review. It is now complete. I raised and issue here: https://github.com/matbesancon/BilevelOptimization.jl/issues/14 regarding documentation.
