### 2019-04-05 70ee1d3
completed docstring, README
