### 2019-04-05 matbesancon/BilevelOptimization.jl#14 @vissarion
The issue is solved by https://github.com/matbesancon/BilevelOptimization.jl/pull/15
