### 2019-04-05 #1278 @vissarion
@matbesancon thanks for addressing the issue on documentation. I closed https://github.com/matbesancon/BilevelOptimization.jl/issues/14 and finished my review check-list.
