### 2019-04-06 #1278 @matbesancon
@vissarion thanks for the discussion, I will adapt the paper to also include the signatures of `build_blp_model`
