### 2019-04-06 0d33fe5
extended doc, developed paper
