+++ src/bilevel_flows.jl
@@ -40,8 +40,8 @@ end
 
 """
 	`build_blp_model(bfp::BilevelFlowProblem, solver; comp_method)`
-	
-Build the JuMP model from the data of `bfp` and assign it the passed solver
+
+Build the JuMP model from the data of `bfp` and assign it the passed solver.
 Return `(m, r, y, f, λ, s)`
 - `m`: `JuMP.Model`
 - `y[i,j,k]`: binary variables indicating which option tax option is chosen by the upper-level
