### 2019-04-06 e965a44
jump model consistency
