### 2019-04-06 38898c6
extended doc, developed paper (#16)
* extended doc, developed paper

* bullet points

* added signature

* jump model consistency

* rephrase and improve

* slack variable
