+++ paper/paper.md
@@ -30,7 +30,7 @@ a set of constraints on the decision.
 Bilevel optimization is a class of mathematical optimization problems
 with the optimality conditions of a lower-level problem embedded in the
 constraints. BilevelOptimization.jl is a toolbox built on top of the JuMP.jl
-modeling package [@dunning2017jump].
+ecosystem for mathematical optimization [@dunning2017jump].
 Bilevel optimization is used to tackle various problems in areas such as
 power systems, security applications, network design or market equilibria.
 See [@dempe2018bilevel] for an overview of applications and recent
@@ -40,8 +40,8 @@ The computation of an optimal solution to a bilevel problem is in general hard.
 Even with all the constraints and the objectives at the two level being linear,
 the resulting problem is non-convex and NP-hard, with possibly a disjoint
 feasible set. Optimization practitioners often rely on problem-specific
-properties and modeling techniques or heuristics, the goal of this package
-is to offer a both flexible model of a general class of bilevel problems
+properties and modeling techniques or heuristics. The goal of this package
+is to offer both a flexible model for a general class of bilevel problems
 and a solving method which is compatible with the JuMP workflow.
 
 # Bilevel optimization
@@ -95,7 +95,8 @@ s.t. G x + H y <= q
 
 The last equation is a complementarity constraint, corresponding
 to the fact that at least one of $(\lambda_i, s_i)$ has to be equal
-to zero. This non-convex, non-linear constraint cannot be tackled
+to zero, where $s_i$ is the slack variable associated with constraint $i$.
+This non-convex, non-linear constraint cannot be tackled
 efficiently by common optimization solvers and needs to be re-formulated.
 The two common approaches are linearization using a binary variable and
 "big-M" primal and dual upper bounds [@pineda19] and Special Ordered Sets
@@ -110,12 +111,30 @@ forcing at least one of them to 0.
 # Types and methods for bilevel optimization
 
 The data for a bilevel linear problems are stored in the `BilevelLP` structure,
-with the same notation as used above. A JuMP `Model` can be built from scratch
+with the same notation as used above. A `JuMP.Model` can be built from scratch
 from these data or passed on to the `build_blp_model` function which adds the
-lower-level feasibility and optimality constraints. The signature of this
-function includes a keyword argument `comp_method` for the choice of method
-to tackle complementarity constraints. The two methods mentioned in the previous
-section are represented as types, `SOS1Complementarity` and `BoundComplementarity`.
+lower-level feasibility and optimality constraints. This function supports different
+signatures through multiple dispatch. To build the model from scratch using
+the data from `BilevelLP`, the function is called as follows:
+```julia
+build_blp_model(bp::BilevelLP, solver; [comp_method])
+```
+
+It returns a tuple `(m, x, y, lambda, s)` with:
+
+- `m` the JuMP model
+- `x` the upper-level vector of variables
+- `y` the lower-level vector of variables
+- `lambda` the lower-level dual variables
+- `s` the lower-level slack variables
+
+Other signatures allow users to modify an already-existing JuMP model
+without storing every constraint or variable into a `BilevelLP` structure.
+
+All signatures of the `build_blp_model` function include a keyword argument
+`comp_method` for the choice of method to tackle complementarity constraints.
+The two methods mentioned in the previous section are represented as types,
+`SOS1Complementarity` and `BoundComplementarity`.
 For the second option, primal and dual bounds can either be scalars or vectors
 to use bounds adapted to each constraint.
 Some problem-specific methods are often used in the literature to handle
@@ -134,7 +153,7 @@ The toll-setting problem is a class of bilevel optimization where the two
 levels of decision are taken on a graph [@brotcorne2001bilevel].
 It belongs to the more general framework of network pricing problems with
[…]
