### 2019-05-12 #1278 @HaoZeke
@labarba, having taken a second look, I believe all previous concerns raised were addressed and I would recommend that this be accepted.
