### 2019-05-12 #1278 @vissarion
@poulson I recommend acceptance. However, I am not clicking the box because, as far as I understand, there are no performance claims by the submitted software to confirm so the box should be left unchecked.
