### 2019-06-03 #1278 @poulson
@matbesancon Would you mind fixing the following typo/grammatical/citation/clarification issues?

"at the two level" -> "at the two levels"

"with possibly a disjoint feasible set" -> "with a possibly disjoint feasible set"

"solving method" -> "solver"

"for such case" -> "for such cases"

"the two principal being" -> "with the two primary approaches being"
[Citations needed for what the two approaches roughly are.]

Clarification: "into a regular one"?
[Citation of Special Ordered Sets of type 1]

"forcing at least one of them to 0" -> "forcing at least one of them to zero"

"m the JuMP model" -> "m: the JuMP model,", etc., adding an "and" in the second-to-last item.

"Where CM is the complementarity" -> "where CM is the..."

"Each arc has an initial cost, the lower-level" ->
  "Each arc has an initial cost, and the lower-level"

"This problem can entirely be modeled using the framework presented above, a" ->
"This problem can be entirely modeled using the framework presented above, using a composite datatype defined in the package for holding all required data,"

Add commas and an "and" in the final list, such as: "The initial matrix of arc costs `init_cost`,"

"Even though BilevelOptimization.jl is thought for" -> "Even though BilevelOptimization.jl is typically for"

After doing so, could you register a DOI on zenodo and post it here?
