### 2019-06-18 #1278 @matbesancon
apologies to all for the delays! I'll edit the paper with @poulson's comments
