### 2019-06-18 bad9c74
Project version update
