+++ Project.toml
@@ -1,7 +1,7 @@
 name = "BilevelOptimization"
 uuid = "98803d92-2a2a-5e5c-9642-fb423c87776e"
 authors = ["Mathieu Besançon <mathieu.besancon@gmail.com>"]
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
