+++ Project.toml
@@ -16,4 +16,5 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 test = ["Test", "Cbc"]
 
 [compat]
+julia = "1"
 JuMP = "0.18"
