### 2019-06-18 #1278 @matbesancon
I'm registering a new version to trigger the DOI creation from Zenodo
