### 2019-06-18 #1278 @matbesancon
The version of the software is v0.2.1, the corresponding DOI is 10.5281/zenodo.3249188
