### 2019-06-18 889c8ff
citation document and DOI
