+++ CITATION.bib
@@ -1,0 +1,8 @@
+@misc{BilevelOptimization.jl,
+  author       = {Mathieu Besançon},
+  title        = {matbesancon/BilevelOptimization.jl},
+  month        = jun,
+  year         = 2019,
+  doi          = {10.5281/zenodo.3249187},
+  url          = {https://doi.org/10.5281/zenodo.3249187}
+}
