### 2019-07-01 #1278 @kyleniemeyer
Hi @matbesancon, a few little things to fix before we publish:

- In your Zenodo archive, the summary/description says "A changelog is being generated, please check back in a few minutes."
- I made a couple of copy-edits to the paper in https://github.com/matbesancon/BilevelOptimization.jl/pull/20, could you merge that?
