### 2019-07-01 #1278 @matbesancon
hi @kyleniemeyer thanks, I've updated the Zenodo description field, I responded on the repo for the edit, I'm not sure about the square bracket thing
