### 2019-07-01 #1278 @kyleniemeyer
Congrats @matbesancon on your submission's publication in JOSS! Thanks to @HaoZeke and @vissarion for reviewing, and @poulson for editing.
